subtogramLines prototype
