+++ app/assets/javascripts/subtogram.js
@@ -1,11 +1,8 @@
-var Subtogram = function(map, style, lines){
-  this.map = map;
-  this.style = style;
+var Subtogram = function(args){
+  args = args || {};
+  this.map = args.map;
+  this.style = args.style;
   this.addLayers();
-
-  for (var line in lines) {
-    if (lines[line].show) this.linesShown.push(line);
-  };
 };
 
 Subtogram.prototype = {
@@ -13,22 +10,9 @@ Subtogram.prototype = {
   style: null,
 
   currentHoverId: {sections: ['none'], stations: ['none']},
-  currentYear: null,
   linesShown: [],
 
-  layers: {
-    sections: {
-      BUILDSTART: 'sections_buildstart',
-      OPENGING: 'sections_opening',
-      HOVER: 'sections_hover'
-    },
-    stations: {
-      BUILDSTART: 'stations_buildstart',
-      OPENGING: 'stations_opening',
-      HOVER: 'stations_hover',
-      INNER_LAYER: 'stations_inner_layer'
-    }
-  },
+  layers: null,
 
   _addSource: function(type) {
     var sourceName = type + '_source';
@@ -69,11 +53,6 @@ Subtogram.prototype = {
     this.map.addLayer(layer);
   },
 
-  setYear: function(year) {
-    this.currentYear = year;
-    this.filter();
-  },
-
   setHoverIds: function(type, ids) {
     if ((ids && this.currentHoverId[type] == ids) ||
         (!ids && this.currentHoverId[type] == ['none'])) return;
@@ -95,51 +74,7 @@ Subtogram.prototype = {
     }
     this.filter();
     return this.linesShown;
-  },
-
-  filter: function() {
-    var self = this;
-
-    var hoverId = this.currentHoverId;
-    var year = this.currentYear;
-
-    ['sections', 'stations'].forEach(function(type){
-      for (var k in self.layers[type]) {
-        var layer = self.layers[type][k];
-        var filter;
-
-        if (layer.indexOf('hover') !== -1){
-          var ids = ["in", "id"].concat(hoverId[type]);
-          filter = ["all", ids];
-        } else if (layer.indexOf('buildstart') !== -1) {
-          filter = [
-            "all",
-            ["<=", "buildstart", year],
-            [">", "buildstart_end", year],
-          ];
-        } else if (layer.indexOf('opening') !== -1){
-          filter = [
-            "all",
-            ["<=", "opening", year],
-            [">", "closure", year],
-          ];
-        } else if (layer.indexOf('inner') !== -1){
-          filter = [
-            "all",
-            ["<=", "buildstart", year],
-            [">", "closure", year],
-          ];
-        }
-
-        if (self.linesShown) {
-          var linesShownFilter = ["in", "line"].concat(self.linesShown);
-          filter.push(linesShownFilter);
-        }
-
-        if (filter) self.map.setFilter(layer, filter);
-      }
-    });
-  },
+  }
 }
 
 module.exports = Subtogram;
